# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Api",
    "version": "1.0.0",
    "description": "API"
  },
  "servers": [
    {
      "url": "https://example/api/"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Вход в систему",
        "description": "Аутентификация пользователя по email и паролю",
        "operationId": "login",
        "requestBody": {
          "description": "Данные для входа",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "example": {
                  "status": "success",
                  "data": {
                    "userId": "4",
                    "tokens": {
                      "accessToken": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                      "refreshToken": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    }
                  },
                  "errors": []
                }
              }
            }
          },
          "401": {
            "description": "Неверные учетные данные",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Регистрация нового пользователя",
        "description": "Создание нового аккаунта пользователя",
        "operationId": "register",
        "requestBody": {
          "description": "Данные для регистрации",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная регистрация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "409": {
            "description": "Пользователь уже существует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Обновление токенов",
        "description": "Обновление access токена с помощью refresh токена",
        "operationId": "refresh",
        "requestBody": {
          "description": "Refresh токен",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "Refresh token",
                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Токены успешно обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                },
                "example": {
                  "status": "success",
                  "data": {
                    "tokens": {
                      "accessToken": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9...",
                      "refreshToken": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                    }
                  },
                  "errors": []
                }
              }
            }
          },
          "401": {
            "description": "Недействительный или отозванный refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Выход из системы",
        "description": "Завершение сессии пользователя",
        "operationId": "logout",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Refresh токен для отзыва (опционально)",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "Refresh token для отзыва"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный выход",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "status": "success",
                  "data": [],
                  "errors": []
                }
              }
            }
          },
          "401": {
            "description": "Требуется аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/password/forgot": {
      "post": {
        "tags": ["Password"],
        "summary": "Запрос на сброс пароля",
        "description": "Отправка кода подтверждения на email для сброса пароля",
        "operationId": "forgotPassword",
        "requestBody": {
          "description": "Email пользователя",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email пользователя",
                    "example": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Код отправлен на email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/password/check": {
      "post": {
        "tags": ["Password"],
        "summary": "Проверка кода сброса пароля",
        "description": "Проверка кода подтверждения для сброса пароля",
        "operationId": "checkCode",
        "requestBody": {
          "description": "Email и код подтверждения",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "code"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email пользователя"
                  },
                  "code": {
                    "type": "string",
                    "description": "Код подтверждения",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Код подтвержден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверный код",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/password/reset": {
      "post": {
        "tags": ["Password"],
        "summary": "Сброс пароля",
        "description": "Установка нового пароля после подтверждения кода",
        "operationId": "resetPassword",
        "requestBody": {
          "description": "Данные для сброса пароля",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "code", "newPassword"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email пользователя"
                  },
                  "code": {
                    "type": "string",
                    "description": "Код подтверждения"
                  },
                  "newPassword": {
                    "type": "string",
                    "format": "password",
                    "description": "Новый пароль",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пароль успешно изменен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/info": {
      "get": {
        "tags": ["User"],
        "summary": "Получение информации о пользователе",
        "description": "Получение профиля текущего авторизованного пользователя",
        "operationId": "getUserInfo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                },
                "example": {
                  "status": "success",
                  "data": {
                    "name": "Иван Иванов",
                    "email": "test3@test.ru",
                    "phone": "89991234567"
                  },
                  "errors": []
                }
              }
            }
          },
          "401": {
            "description": "Требуется аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User"],
        "summary": "Изменение информации пользователя",
        "description": "Обновление профиля текущего авторизованного пользователя",
        "operationId": "updateUserInfo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Данные для обновления",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Имя пользователя"
                  },
                  "phone": {
                    "type": "string",
                    "description": "Номер телефона"
                  },
                  "company": {
                    "type": "string",
                    "description": "Компания"
                  },
                  "post": {
                    "type": "string",
                    "description": "Должность"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Информация успешно обновлена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Требуется аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Пароль",
            "example": "your_password"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "fullName"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "fullName": {
            "type": "string",
            "description": "Полное имя (Фамилия Имя Отчество)",
            "example": "Иванов Иван Иванович"
          },
          "phone": {
            "type": "string",
            "description": "Номер телефона",
            "example": "+79991234567"
          },
          "company": {
            "type": "string",
            "description": "Компания",
            "example": "ООО Ромашка"
          },
          "post": {
            "type": "string",
            "description": "Должность",
            "example": "Менеджер"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "error"],
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "string",
                "description": "ID пользователя"
              },
              "tokens": {
                "type": "object",
                "properties": {
                  "accessToken": {
                    "type": "string",
                    "description": "Access token для доступа к API"
                  },
                  "refreshToken": {
                    "type": "string",
                    "description": "Refresh token для обновления access token"
                  }
                }
              }
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": []
          }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "error"],
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "tokens": {
                "type": "object",
                "properties": {
                  "accessToken": {
                    "type": "string",
                    "description": "Новый access token"
                  },
                  "refreshToken": {
                    "type": "string",
                    "description": "Новый refresh token"
                  }
                }
              }
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": []
          }
        }
      },
      "UserInfoResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "error"],
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "Имя пользователя"
              },
              "email": {
                "type": "string",
                "format": "email",
                "description": "Email пользователя"
              },
              "phone": {
                "type": "string",
                "description": "Номер телефона"
              },
              "company": {
                "type": "string",
                "description": "Компания"
              },
              "post": {
                "type": "string",
                "description": "Должность"
              }
            }
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": []
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success", "error"],
            "example": "success"
          },
          "data": {
            "type": "array",
            "items": {},
            "description": "Пустой массив или дополнительные данные",
            "example": []
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": []
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"],
            "example": "error"
          },
          "data": {
            "type": "array",
            "items": {},
            "example": []
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Неверные учетные данные"]
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Вставьте access token в формате: Bearer {token}"
      }
    }
  }
}